+++ src/views/MainPage/components/Introduce/style.ts
@@ -62,6 +62,20 @@ export const TextContainer = styled(motion.span)`
   }
 `;
 
+export const MotionTitle = styled(motion.span)`
+  position: absolute;
+  left: 0;
+
+  :before {
+    content: attr(data-text);
+    display: inline-block;
+    background: linear-gradient(93deg, #c9d8dd 78.65%, #fff 128.82%, #c9d8dd 137.19%);
+    background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+  }
+`;
+
 export const BackgroundTitle = styled.span`
   background: linear-gradient(93deg, #c9d8dd -34.77%, #fff -14.77%, #636770 6.12%);
   background-clip: text;
